Updated code to include SPI comms to onboard IMU.
This IMU will be used primarily for testing of the algorithm. If it
works well, it may be used in tandem with MPU-6050. The more sensors the
merrier!

Updated the hall effect sensor configurations to use defines to minimize
errors when changing which timers they're apart of. Using TIM2 and TIM1
now instead of TIM3 as these are the only timers with pinouts. Most
likely won't be able to use DMA for any of them.

Signed-off-by: Ethan Tola <ext9285@g.rit.edu>