Added built-in imus to the sequence.
Running it again, its obvious that the onboard imu's are either damaged
(the accelerometer doesn't response to stimuli) or incredibly noisy that
no error correction will fix it without losing actual data. Regardless,
they will be left in as to test with a newer board that may not be as
damaged.

Signed-off-by: Ethan Tola <ethanltola@gmail.com>